+++ jjy3385/tiny/cgi-bin/adder.c
@@ -4,15 +4,13 @@
 /* $begin adder */
 #include "csapp.h"
 
-int main(void)
-{
+int main(void) {
   char *buf, *p;
   char arg1[MAXLINE], arg2[MAXLINE], content[MAXLINE];
   int n1 = 0, n2 = 0;
 
   /* Extract the two arguments */
-  if ((buf = getenv("QUERY_STRING")) != NULL)
-  {
+  if ((buf = getenv("QUERY_STRING")) != NULL) {
     p = strchr(buf, '&');
     *p = '\0';
     strcpy(arg1, buf);
@@ -25,8 +23,7 @@ int main(void)
   sprintf(content, "QUERY_STRING=%s\r\n<p>", buf);
   sprintf(content + strlen(content), "Welcome to add.com: ");
   sprintf(content + strlen(content), "THE Internet addition portal.\r\n<p>");
-  sprintf(content + strlen(content), "The answer is: %d + %d = %d\r\n<p>",
-          n1, n2, n1 + n2);
+  sprintf(content + strlen(content), "The answer is: %d + %d = %d\r\n<p>", n1, n2, n1 + n2);
   sprintf(content + strlen(content), "Thanks for visiting!\r\n");
 
   /* Generate the HTTP response */
