+++ jjy3385/tiny/csapp.h
@@ -4,7 +4,6 @@
 /* $begin csapp.h */
 #ifndef __CSAPP_H__
 #define __CSAPP_H__
-
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdarg.h>
