+++ jjy3385/tiny/home.html
@@ -1,7 +1,16 @@
 <html>
-<head><title>test</title></head>
-<body> 
-<img align="middle" src="godzilla.gif">
-Dave O'Hallaron
+
+<head>
+    <title>tiny</title>
+    <link rel="icon" href="data:,">
+</head>
+
+<body>
+    <video autoplay muted loop controls>
+        <source src="burning_godzilla.mp4" type="video/mp4">
+    </video>    
+    <img src="godzilla.gif">
+    나만의 TINY 서버
 </body>
+
 </html>
