+++ jjy3385/tiny/tiny.c
@@ -17,23 +17,20 @@ void serve_dynamic(int fd, char *filename, char *cgiargs);
 void clienterror(int fd, char *cause, char *errnum, char *shortmsg,
                  char *longmsg);
 
-int main(int argc, char **argv)
-{
+int main(int argc, char **argv) {
   int listenfd, connfd;
   char hostname[MAXLINE], port[MAXLINE];
   socklen_t clientlen;
   struct sockaddr_storage clientaddr;
 
   /* Check command line args */
-  if (argc != 2)
-  {
+  if (argc != 2) {
     fprintf(stderr, "usage: %s <port>\n", argv[0]);
     exit(1);
   }
 
   listenfd = Open_listenfd(argv[1]);
-  while (1)
-  {
+  while (1) {
     clientlen = sizeof(clientaddr);
     connfd = Accept(listenfd, (SA *)&clientaddr,
                     &clientlen); // line:netp:tiny:accept
@@ -43,4 +40,238 @@ int main(int argc, char **argv)
     doit(connfd);  // line:netp:tiny:doit
     Close(connfd); // line:netp:tiny:close
   }
+}
+
+/*
+ * doit - handle one HTTP request/response transaction
+ */
+void doit(int fd) {
+  int is_static;
+  struct stat sbuf;
+  char buf[MAXLINE], method[MAXLINE], uri[MAXLINE], version[MAXLINE];
+  char filename[MAXLINE], cgiargs[MAXLINE];
+  rio_t rio;
+
+  /* Read request line and headers */
+  Rio_readinitb(&rio, fd);
+  if (!Rio_readlineb(&rio, buf, MAXLINE))
+    return;
+  printf("%s", buf);
+  sscanf(buf, "%s %s %s", method, uri, version);
+  if (strcasecmp(method, "GET")) {
+    clienterror(fd, method, "501", "Not Implemented",
+                "Tiny does not implement this method");
+    return;
+  }
+  read_requesthdrs(&rio);
+
+  /* Parse URI from GET request */
+  is_static = parse_uri(uri, filename, cgiargs);
+  if (stat(filename, &sbuf) < 0) {
+    clienterror(fd, filename, "404", "Not found",
+                "Tiny couldn't find this file");
+    return;
+  }
+
+  if (is_static) { /* Serve static content */
+    if (!(S_ISREG(sbuf.st_mode)) || !(S_IRUSR & sbuf.st_mode)) {
+      clienterror(fd, filename, "403", "Forbidden",
+                  "Tiny couldn't read the file");
+      return;
+    }
+    serve_static(fd, filename, sbuf.st_size);
+  } else { /* Serve dynamic content */
+    if (!(S_ISREG(sbuf.st_mode)) || !(S_IXUSR & sbuf.st_mode)) {
+      clienterror(fd, filename, "403", "Forbidden",
+                  "Tiny couldn't run the CGI program");
+      return;
+    }
+    serve_dynamic(fd, filename, cgiargs);
+  }
+}
+
+/*
+ * read_requesthdrs - read HTTP request headers
+ */
+void read_requesthdrs(rio_t *rp) {
+  char buf[MAXLINE];
+
+  Rio_readlineb(rp, buf, MAXLINE);
+  printf("%s", buf);
+  while (strcmp(buf, "\r\n")) {
+    Rio_readlineb(rp, buf, MAXLINE);
+    printf("%s", buf);
+  }
+  return;
+}
+
+/*
+ * parse_uri - parse URI into filename and CGI args
+ *             return 0 if dynamic content, 1 if static
+ */
+int parse_uri(char *uri, char *filename, char *cgiargs) {
+  char *ptr;
+
+  if (!strstr(uri, "cgi-bin")) { /* Static content */
+    strcpy(cgiargs, "");
+    strcpy(filename, ".");
+    strcat(filename, uri);
+    if (uri[strlen(uri) - 1] == '/')
+      strcat(filename, "home.html");
+    return 1;
+  } else { /* Dynamic content */
+    ptr = index(uri, '?');
+    if (ptr) {
+      strcpy(cgiargs, ptr + 1);
+      *ptr = '\0';
+    } else
+      strcpy(cgiargs, "");
+    strcpy(filename, ".");
+    strcat(filename, uri);
+    return 0;
+  }
+}
+
+/*
+ * serve_static - copy a file back to the client
+ */
+void serve_static(int fd, char *filename, int filesize) {
+  int srcfd;
+  char *srcp, filetype[MAXLINE];
+
+  char buf[MAXBUF];
+  char *p = buf;
+  int n;
+  int remaining = sizeof(buf);
+
+  /* Send response headers to client */
+  get_filetype(filename, filetype);
+
+  /* Build the HTTP response headers correctly - use separate buffers or append
+   */
+  n = snprintf(p, remaining, "HTTP/1.0 200 OK\r\n");
+  p += n;
+  remaining -= n;
+
+  n = snprintf(p, remaining, "Server: Tiny Web Server\r\n");
+  p += n;
+  remaining -= n;
+
+  n = snprintf(p, remaining, "Connection: close\r\n");
+  p += n;
+  remaining -= n;
+
+  n = snprintf(p, remaining, "Content-length: %d\r\n", filesize);
+  p += n;
+  remaining -= n;
+
+  n = snprintf(p, remaining, "Content-type: %s\r\n\r\n", filetype);
+  p += n;
+  remaining -= n;
+
+  Rio_writen(fd, buf, strlen(buf));
+  printf("Response headers:\n");
+  printf("%s", buf);
+
+  /* Send response body to client */
+  srcfd = Open(filename, O_RDONLY, 0);
+  srcp = Mmap(0, filesize, PROT_READ, MAP_PRIVATE, srcfd, 0);
+  Close(srcfd);
+  Rio_writen(fd, srcp, filesize);
+  Munmap(srcp, filesize);
+}
+
+/*
+ * get_filetype - derive file type from file name
+ */
+void get_filetype(char *filename, char *filetype) {
+  if (strstr(filename, ".html"))
+    strcpy(filetype, "text/html");
+  else if (strstr(filename, ".gif"))
+    strcpy(filetype, "image/gif");
+  else if (strstr(filename, ".png"))
+    strcpy(filetype, "image/png");
+  else if (strstr(filename, ".jpg"))
+    strcpy(filetype, "image/jpeg");
+  else if (strstr(filename, ".mp4"))
+    strcpy(filetype, "video/mp4");
+  else
+    strcpy(filetype, "text/plain");
+}
+
+/*
+ * serve_dynamic - run a CGI program on behalf of the client
+ */
+void serve_dynamic(int fd, char *filename, char *cgiargs) {
+  char buf[MAXLINE], *emptylist[] = {NULL};
+  pid_t pid;
+
+  /* Return first part of HTTP response */
+  sprintf(buf, "HTTP/1.0 200 OK\r\n");
+  Rio_writen(fd, buf, strlen(buf));
+  sprintf(buf, "Server: Tiny Web Server\r\n");
+  Rio_writen(fd, buf, strlen(buf));
+
+  /* Create a child process to handle the CGI program */
+  if ((pid = Fork()) < 0) { /* Fork failed */
+    perror("Fork failed");
+    return;
+  }
+
+  if (pid == 0) { /* Child process */
+    /* Real server would set all CGI vars here */
+    setenv("QUERY_STRING", cgiargs, 1);
+
+    /* Redirect stdout to client */
+    if (Dup2(fd, STDOUT_FILENO) < 0) {
+      perror("Dup2 error");
+      exit(1);
+    }
+    Close(fd);
+
+    /* Run CGI program */
+    Execve(filename, emptylist, environ);
+
+    /* If we get here, Execve failed */
+    perror("Execve error");
+    exit(1);
+  } else { /* Parent process */
+    /* Parent waits for child to terminate */
+    int status;
+    if (waitpid(pid, &status, 0) < 0) {
+      perror("Wait error");
+    }
+
+    printf("Child process %d terminated with status %d\n", pid, status);
+    /* Parent continues normally - returns to doit() */
+  }
+  /* When we return from here, doit() will close the connection */
+}
+
+/*
+ * clienterror - returns an error message to the client
+ */
+void clienterror(int fd, char *cause, char *errnum, char *shortmsg,
+                 char *longmsg) {
+  char buf[MAXLINE], body[MAXBUF];
+
+  /* Build the HTTP response body */
+  sprintf(body, "<html><title>Tiny Error</title>");
+  sprintf(body,
+          "%s<body bgcolor="
+          "ffffff"
+          ">\r\n",
+          body);
+  sprintf(body, "%s%s: %s\r\n", body, errnum, shortmsg);
+  sprintf(body, "%s<p>%s: %s\r\n", body, longmsg, cause);
+  sprintf(body, "%s<hr><em>The Tiny Web server</em>\r\n", body);
+
+  /* Print the HTTP response */
+  sprintf(buf, "HTTP/1.0 %s %s\r\n", errnum, shortmsg);
+  Rio_writen(fd, buf, strlen(buf));
+  sprintf(buf, "Content-type: text/html\r\n");
+  Rio_writen(fd, buf, strlen(buf));
+  sprintf(buf, "Content-length: %d\r\n\r\n", (int)strlen(body));
+  Rio_writen(fd, buf, strlen(buf));
+  Rio_writen(fd, body, strlen(body));
 }
